Clean up and custom build
