+++ package.json
@@ -8,23 +8,16 @@
     "url": "https://github.com/mozilla/notes/issues"
   },
   "dependencies": {
-    "@ckeditor/ckeditor5-adapter-ckfinder": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-autoformat": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-basic-styles": "^1.0.0-alpha.1",
-    "@ckeditor/ckeditor5-block-quote": "^1.0.0-alpha.1",
-    "@ckeditor/ckeditor5-build-classic": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-dev-utils": "^4.0.1",
     "@ckeditor/ckeditor5-dev-webpack-plugin": "^2.0.22",
-    "@ckeditor/ckeditor5-easy-image": "^1.0.0-alpha.1",
-    "@ckeditor/ckeditor5-editor-classic": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-essentials": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-heading": "^1.0.0-alpha.1",
-    "@ckeditor/ckeditor5-image": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-link": "^1.0.0-alpha.1",
     "@ckeditor/ckeditor5-list": "^1.0.0-alpha.1",
-    "@ckeditor/ckeditor5-markdown-gfm": "^1.0.0-alpha.1",
+    "@ckeditor/ckeditor5-markdown-gfm": "^1.0.0-alpha.2",
     "@ckeditor/ckeditor5-paragraph": "^1.0.0-alpha.1",
-    "@ckeditor/ckeditor5-upload": "^1.0.0-alpha.1",
     "babel-minify-webpack-plugin": "^0.2.0",
     "css-loader": "^0.28.7",
     "node-sass": "^4.6.1",
@@ -50,6 +43,7 @@
   "repository": "mozilla/notes",
   "scripts": {
     "build": "node scripts/build-locales && web-ext build -s src",
+    "build-ck": "cd ckeditor-build && webpack && cd ..",
     "clean": "rimraf web-ext-artifacts 'src/sidebar/vendor/!(.gitkeep)' && npm run postinstall",
     "format": "prettier 'src/**/!(vendor)/*.{css,js}' --single-quote --write",
     "lint": "npm-run-all lint:*",
@@ -57,7 +51,7 @@
     "lint:webext": "web-ext lint -s src --ignore-files='sidebar/vendor/*.js' --self-hosted",
     "package": "npm run build && mv web-ext-artifacts/*.zip addon.xpi",
     "postformat": "npm run lint:js -- --fix",
-    "postinstall": "node scripts/postinstall.js && node scripts/build-locales.js",
+    "postinstall": "npm run build-ck && node scripts/postinstall.js && node scripts/build-locales.js",
     "prebuild": "npm run clean",
     "start": "web-ext run -s src",
     "start-rtl": "WEB_EXT_PREF=general.useragent.locale=ar npm start",
